+++ CHANGELOG.md
@@ -13,6 +13,13 @@ major.minor[.revision]
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+## 0.1.4
+
+2021-09-21
+
+- Assure PHP 7.3 compatibility.
+- Add `$returnObject` argument to the `validateSignedRequestData` function.
+
 ## 0.1.3
 
 2021-09-19